move to lib and resolve deprecation warnigns
