+++ test/tools/parser_throughput/main.cc
@@ -4,9 +4,9 @@
 #include "vw/config/options_cli.h"
 #include "vw/core/io_buf.h"
 #include "vw/core/learner.h"
-#include "vw/core/parse_example_json.h"
 #include "vw/core/vw.h"
 #include "vw/io/io_adapter.h"
+#include "vw/json_parser/parse_example_json.h"
 #include "vw/text_parser/parse_example_text.h"
 
 #ifdef VW_BUILD_CSV
@@ -172,7 +172,7 @@ int main(int argc, char** argv)
     {
       VW::multi_ex examples;
       examples.push_back(&VW::get_unused_example(vw));
-      VW::read_line_decision_service_json<false>(*vw, examples, const_cast<char*>(line.data()), line.length(), false,
+      VW::parsers::json::read_line_decision_service_json<false>(*vw, examples, const_cast<char*>(line.data()), line.length(), false,
           (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &interaction);
       VW::finish_example(*vw, examples);
     }
