+++ test/unit_test/CMakeLists.txt
@@ -21,14 +21,12 @@ add_executable(vw-unit-test.out
   continuous_actions_parser_test.cc
   custom_reduction_test.cc
   distributionally_robust_test.cc
-  dsjson_parser_test.cc
   epsilon_test.cc
   example_header_test.cc
   example_test.cc
   feature_group_test.cc
   guard_test.cc
   interactions_test.cc
-  json_parser_test.cc
   loss_functions_test.cc
   main.cc
   math_test.cc
