+++ vowpalwabbit/CMakeLists.txt
@@ -13,6 +13,7 @@ if(VW_BUILD_CSV)
 endif()
 add_subdirectory(explore)
 add_subdirectory(io)
+add_subdirectory(json_parser)
 add_subdirectory(model_merger)
 add_subdirectory(slim)
 add_subdirectory(spanning_tree_bin)
