+++ vowpalwabbit/json_parser/include/vw/json_parser/parse_example_slates_json.h
@@ -1,38 +1,36 @@
 #pragma once
 
+#include "json_utils.h"
 #include "vw/common/future_compat.h"
-#include "vw/json_parser/parse_example_slates_json.h"
+#include "vw/core/example.h"
+#include "vw/core/hashstring.h"
+#include "vw/core/label_parser.h"
 
+#include <unordered_map>
+
+namespace VW
+{
+namespace parsers
+{
+namespace json
+{
+namespace details
+{
 template <bool audit>
-VW_DEPRECATED("parse_slates_example_json moved to VW::parsers::json::details::parse_slates_example_json")
 void parse_slates_example_json(const VW::label_parser& lbl_parser, hash_func_t hash_func, uint64_t hash_seed,
     uint64_t parse_mask, bool chain_hash, VW::multi_ex& examples, char* line, size_t length,
     VW::example_factory_t example_factory, void* ex_factory_context,
-    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr)
-{
-  VW::parsers::json::details::parse_slates_example_json<audit>(lbl_parser, hash_func, hash_seed, parse_mask, chain_hash,
-      examples, line, length, example_factory, ex_factory_context, dedup_examples);
-}
+    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr);
 
 template <bool audit>
-VW_DEPRECATED("parse_slates_example_json moved to VW::parsers::json::details::parse_slates_example_json")
 void parse_slates_example_json(const VW::workspace& all, VW::multi_ex& examples, char* line, size_t length,
     VW::example_factory_t example_factory, void* ex_factory_context,
-    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr)
-{
-  VW::parsers::json::details::parse_slates_example_json<audit>(
-      all, examples, line, length, example_factory, ex_factory_context, dedup_examples);
-}
+    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr);
 
 template <bool audit>
-VW_DEPRECATED("parse_slates_example_dsjson moved to VW::parsers::json::details::parse_slates_example_dsjson")
 void parse_slates_example_dsjson(VW::workspace& all, VW::multi_ex& examples, char* line, size_t length,
     VW::example_factory_t example_factory, void* ex_factory_context, VW::details::decision_service_interaction* data,
-    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr)
-{
-  VW::parsers::json::details::parse_slates_example_dsjson<audit>(
-      all, examples, line, length, example_factory, ex_factory_context, data, dedup_examples);
-}
+    std::unordered_map<uint64_t, VW::example*>* dedup_examples = nullptr);
 
 // Define extern template specializations so they don't get initialized when this file is included
 extern template void parse_slates_example_json<true>(const VW::label_parser& lbl_parser, hash_func_t hash_func,
@@ -57,3 +55,7 @@ extern template void parse_slates_example_dsjson<true>(VW::workspace& all, VW::m
 extern template void parse_slates_example_dsjson<false>(VW::workspace& all, VW::multi_ex& examples, char* line,
     size_t length, VW::example_factory_t example_factory, void* ex_factory_context,
     VW::details::decision_service_interaction* data, std::unordered_map<uint64_t, VW::example*>* dedup_examples);
+}  // namespace details
+}  // namespace json
+}  // namespace parsers
+}  // namespace VW
