+++ vowpalwabbit/json_parser/tests/json_parser_test.cc
@@ -2,13 +2,16 @@
 // individual contributors. All rights reserved. Released under a BSD (revised)
 // license as described in the file LICENSE.
 
-#include "test_common.h"
+#include "vw/test_common/test_common.h"
 #include "vw/core/reductions/conditional_contextual_bandit.h"
 #include "vw/core/vw.h"
 
 #include <boost/test/test_tools.hpp>
 #include <boost/test/unit_test.hpp>
 #include <vector>
+
+constexpr float FLOAT_TOL = 0.0001f;
+
 // TODO: Make unit test dig out and verify features.
 BOOST_AUTO_TEST_CASE(parse_json_simple)
 {
@@ -24,7 +27,7 @@ BOOST_AUTO_TEST_CASE(parse_json_simple)
       }
     })";
 
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 1);
   BOOST_CHECK_CLOSE(examples[0]->l.simple.label, 1.f, FLOAT_TOL);
@@ -49,7 +52,7 @@ BOOST_AUTO_TEST_CASE(parse_json_simple_with_weight)
       }
     })";
 
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 1);
   BOOST_CHECK_CLOSE(examples[0]->l.simple.label, -1.f, FLOAT_TOL);
@@ -89,7 +92,7 @@ BOOST_AUTO_TEST_CASE(parse_json_cb)
     })";
 
   auto vw = VW::initialize("--cb_adf --json --chain_hash --no_stdin --quiet", nullptr, false, nullptr, nullptr);
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
   BOOST_CHECK_EQUAL(examples.size(), 4);
 
   BOOST_CHECK_EQUAL(examples[0]->l.cb.costs.size(), 1);
@@ -133,7 +136,7 @@ BOOST_AUTO_TEST_CASE(parse_json_cats)
   auto vw =
       VW::initialize("--json --chain_hash --cats 4 --min_value=185 --max_value=23959 --bandwidth 1 --no_stdin --quiet",
           nullptr, false, nullptr, nullptr);
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 1);
 
@@ -168,7 +171,7 @@ BOOST_AUTO_TEST_CASE(parse_json_cats_no_label)
   auto vw = VW::initialize(
       "--json --chain_hash -t --cats 4 --min_value=185 --max_value=23959 --bandwidth 1 --no_stdin --quiet", nullptr,
       false, nullptr, nullptr);
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 1);
   BOOST_CHECK_EQUAL(examples[0]->l.cb_cont.costs.size(), 0);
@@ -236,7 +239,7 @@ BOOST_AUTO_TEST_CASE(parse_json_ccb)
   auto vw =
       VW::initialize("--ccb_explore_adf --json --chain_hash --no_stdin --quiet", nullptr, false, nullptr, nullptr);
 
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 8);
   BOOST_CHECK_EQUAL(examples[0]->l.conditional_contextual_bandit.type, VW::ccb_example_type::SHARED);
@@ -305,7 +308,7 @@ BOOST_AUTO_TEST_CASE(parse_json_cb_as_ccb)
   auto vw =
       VW::initialize("--ccb_explore_adf --json --chain_hash --no_stdin --quiet", nullptr, false, nullptr, nullptr);
 
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 5);
   BOOST_CHECK_EQUAL(examples[0]->l.conditional_contextual_bandit.type, VW::ccb_example_type::SHARED);
@@ -380,7 +383,7 @@ BOOST_AUTO_TEST_CASE(parse_json_slates_dom_parser)
   // Assert parsed values against what they should be
   auto slates_vw = VW::initialize(
       "--slates --dsjson --chain_hash --no_stdin --noconstant --quiet", nullptr, false, nullptr, nullptr);
-  auto examples = parse_json(*slates_vw, json_text);
+  auto examples = vwtest::parse_json(*slates_vw, json_text);
 
   BOOST_CHECK_EQUAL(examples.size(), 8);
   BOOST_CHECK_EQUAL(examples[0]->l.slates.type, VW::slates::example_type::SHARED);
@@ -449,7 +452,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -457,7 +460,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -465,7 +468,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb)
 
   // parse json that includes dedup id's and re-use the examples from the dedup map instead of creating new ones
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples);
 
   BOOST_CHECK_EQUAL(examples.size(), 3);                    // shared example + 2 multi examples
@@ -515,7 +518,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb_missing_dedup_id)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -523,7 +526,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb_missing_dedup_id)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -532,7 +535,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_cb_missing_dedup_id)
   // parse json that includes dedup id's and re-use the examples from the dedup map instead of creating new ones
   examples.push_back(&VW::get_unused_example(vw));
   BOOST_REQUIRE_THROW(
-      VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+      VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
           (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples),
       VW::vw_exception);
 
@@ -584,7 +587,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -592,7 +595,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -600,7 +603,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb)
 
   // parse json that includes dedup id's and re-use the examples from the dedup map instead of creating new ones
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples);
 
   BOOST_CHECK_EQUAL(examples.size(), 6);                    // shared example + 2 multi examples + 3 slots
@@ -704,7 +707,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb_dedup_id_missing)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -712,7 +715,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb_dedup_id_missing)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -721,7 +724,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_ccb_dedup_id_missing)
   // parse json that includes dedup id's and re-use the examples from the dedup map instead of creating new ones
   examples.push_back(&VW::get_unused_example(vw));
   BOOST_REQUIRE_THROW(
-      VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+      VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
           (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples),
       VW::vw_exception);
 
@@ -752,7 +755,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -760,7 +763,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -768,7 +771,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates)
 
   // parse json that includes dedup id's and re-use the examples from the dedup map instead of creating new ones
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples);
 
   BOOST_CHECK_EQUAL(examples.size(), 5);                    // shared example + 2 multi examples + 2 slots
@@ -830,7 +833,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates_dedup_id_missing)
 
   // parse first dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_1.c_str(), action_1.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_1, examples[0]);
 
@@ -838,7 +841,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates_dedup_id_missing)
 
   // parse second dedup example and store it in dedup_examples map
   examples.push_back(&VW::get_unused_example(vw));
-  VW::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
+  VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)action_2.c_str(), action_2.length(),
       (VW::example_factory_t)&VW::get_unused_example, (void*)vw);
   dedup_examples.emplace(dedup_id_2, examples[0]);
 
@@ -848,7 +851,7 @@ BOOST_AUTO_TEST_CASE(parse_json_dedup_slates_dedup_id_missing)
   examples.push_back(&VW::get_unused_example(vw));
 
   BOOST_REQUIRE_THROW(
-      VW::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
+      VW::parsers::json::read_line_json_s<true>(*vw, examples, (char*)json_deduped_text.c_str(), json_deduped_text.length(),
           (VW::example_factory_t)&VW::get_unused_example, (void*)vw, &dedup_examples),
       VW::vw_exception);
 
@@ -878,7 +881,7 @@ BOOST_AUTO_TEST_CASE(parse_json_simple_verify_extents)
       }
     })";
 
-  auto examples = parse_json(*vw, json_text);
+  auto examples = vwtest::parse_json(*vw, json_text);
   BOOST_CHECK_EQUAL(examples[0]->feature_space[' '].size(), 1);
   BOOST_CHECK_EQUAL(examples[0]->feature_space['f'].size(), 5);
   BOOST_CHECK_EQUAL(examples[0]->feature_space['n'].size(), 1);
