+++ vowpalwabbit/json_parser/tests/main.cc
@@ -1,0 +1,6 @@
+// Copyright (c) by respective owners including Yahoo!, Microsoft, and
+// individual contributors. All rights reserved. Released under a BSD (revised)
+// license as described in the file LICENSE.
+
+#define BOOST_TEST_MODULE Main
+#include <boost/test/unit_test.hpp>
