+++ vowpalwabbit/test_common/include/vw/test_common/test_common.h
@@ -6,6 +6,8 @@
 
 #include "vw/config/options.h"
 #include "vw/config/options_cli.h"
+#include "vw/core/vw.h"
+#include "vw/json_parser/parse_example_json.h"
 
 #include <vector>
 
@@ -16,4 +18,9 @@ std::unique_ptr<VW::config::options_i> make_args(ArgsT const&... args)
 {
   return std::unique_ptr<VW::config::options_cli>(new VW::config::options_cli(std::vector<std::string>({args...})));
 }
+
+VW::multi_ex parse_json(VW::workspace& all, const std::string& line);
+
+VW::multi_ex parse_dsjson(
+    VW::workspace& all, std::string line, VW::details::decision_service_interaction* interaction = nullptr);
 }  // namespace vwtest
